+++ SoftRendererLib/src/core/Renderers/TransformedTextureRenderer.h
@@ -15,14 +15,14 @@ namespace Renderer2D
         ~TransformedTextureRenderer() = default;
 
 
-        void DrawTexture(Texture &texture, uint16_t x, uint16_t y, float angle, int32_t offsetX = 0, int32_t offsetY = 0);
+        void DrawTexture(Texture &texture, int16_t x, int16_t y, float angle, int16_t offsetX = 0, int16_t offsetY = 0);
         
-        void DrawTexture(Texture &texture, uint16_t x, uint16_t y,
+        void DrawTexture(Texture &texture, int16_t x, int16_t y,
                          float scaleX, float scaleY);
 
-        void DrawTexture(Texture &texture, uint16_t x, uint16_t y,
+        void DrawTexture(Texture &texture, int16_t x, int16_t y,
                          float scaleX, float scaleY, float angle,
-                         int32_t offsetX, int32_t offsetY);
+                         int16_t offsetX, int16_t offsetY);
 
 
     private:
